server/requirements.txt
fastapi
uvicorn
motor
beanie
python-dotenv
